:tractor: re-organize template target parsing functionality for moban file
