+++ moban/mobanfile/targets.py
@@ -1,29 +1,29 @@
 from moban import constants
-from moban.utils import handle_template
+from moban.mobanfile.templates import handle_template
 from moban.definitions import TemplateTarget
 
 
 def parse_targets(options, targets):
     for target in targets:
         if constants.LABEL_OUTPUT in target:
-            for template_target in handle_explicit_target(options, target):
+            for template_target in _handle_explicit_target(options, target):
                 yield template_target
         else:
             for output, template_file in target.items():
                 if isinstance(template_file, str) is False:
                     # grouping by template type feature
-                    for template_target in handle_group_target(
+                    for template_target in _handle_group_target(
                         options, template_file, output
                     ):
                         yield template_target
                 else:
-                    for template_target in handle_implicit_target(
+                    for template_target in _handle_implicit_target(
                         options, template_file, output
                     ):
                         yield template_target
 
 
-def handle_explicit_target(options, target):
+def _handle_explicit_target(options, target):
     common_data_file = options[constants.LABEL_CONFIG]
     default_template_type = options[constants.LABEL_TEMPLATE_TYPE]
     template_file = target.get(
@@ -46,11 +46,11 @@ def handle_explicit_target(options, target):
                 )
 
 
-def handle_group_target(options, template_file, output):
+def _handle_group_target(options, template_file, output):
     # grouping by template type feature
     common_data_file = options[constants.LABEL_CONFIG]
     group_template_type = output
-    for _output, _template_file in iterate_list_of_dicts(template_file):
+    for _output, _template_file in _iterate_list_of_dicts(template_file):
         for src, dest, t_type in handle_template(
             _template_file, _output, options[constants.LABEL_TMPL_DIRS]
         ):
@@ -59,7 +59,7 @@ def handle_group_target(options, template_file, output):
             )
 
 
-def handle_implicit_target(options, template_file, output):
+def _handle_implicit_target(options, template_file, output):
     common_data_file = options[constants.LABEL_CONFIG]
     default_template_type = options[constants.LABEL_TEMPLATE_TYPE]
     for src, dest, t_type in handle_template(
@@ -73,7 +73,7 @@ def handle_implicit_target(options, template_file, output):
             )
 
 
-def iterate_list_of_dicts(list_of_dict):
+def _iterate_list_of_dicts(list_of_dict):
     for adict in list_of_dict:
         for key, value in adict.items():
             yield (key, value)
